requirements.txt
asgiref==3.10.0
Django==4.2
django-ckeditor==6.7.3
django-js-asset==3.1.2
django-rest-framework==0.1.0
djangorestframework==3.16.1
dotenv==0.9.9
pillow==12.0.0
python-dotenv==1.1.1
sqlparse==0.5.3
tzdata==2025.2
